% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_user_for_filename.R,
%   R/ask_user_for_first_header_row.R
\name{ask_user_for_filename}
\alias{ask_user_for_filename}
\title{Get filename}
\usage{
ask_user_for_filename(file_path = ".")

ask_user_for_filename(file_path = ".")
}
\arguments{
\item{file_path}{Character string. If left blank, the working directory is used.}
}
\value{
Character string.

Character string.
}
\description{
Ask the user to select the file they want to work with. All files in the current directory are given as options.

Ask the user to select the file they want to work with. All files in the current directory are given as options.
}
\examples{
ask_user_for_filename()
ask_user_for_filename()
}
