% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_user_for_first_header_row.R
\name{ask_user_for_first_header_row}
\alias{ask_user_for_first_header_row}
\title{Get first header row}
\usage{
ask_user_for_first_header_row(tab_name, row_number)
}
\arguments{
\item{tab_name}{Character string.}

\item{row_numebr}{Numeric.}
}
\value{
Character string.
}
\description{
Ask the user if the default first header row is correct for a particular set
of data. If the answer is no, the user is asked to input the correct name.
}
\examples{
tab_name <- ask_user_for_first_header_row("Table 1", 4)
}
