% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_user_interaction.R
\name{ask_user_for_tab_name}
\alias{ask_user_for_tab_name}
\title{Get tab name}
\usage{
ask_user_for_tab_name(disaggregation, tab_name)
}
\arguments{
\item{disaggregation, tab_name}{Character string.}
}
\value{
Character string.
}
\description{
Ask the user if the default table name is correct for a particular set of data. If the answer is no, the user is asked to input the correct name.
}
\examples{
tab_name <- ask_user_for_tab_name("Region", "Table 1")
}
