% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_calculations.R
\name{calculate_valid_rates_per_1000}
\alias{calculate_valid_rates_per_1000}
\title{Rates per 1000}
\usage{
calculate_valid_rates_per_1000(numerator, denominator, decimal_places)
}
\arguments{
\item{numerator, denominator, decimal_places}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Calculates the rate per 1000, when the numerator is greater than 3.
}
\examples{
calculate_valid_rates_per_1000(13, 11000, 2)
calculate_valid_rates_per_1000(3, 11000, 2)

}
