% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean_strings}
\alias{clean_strings}
\title{clean all character strings in dataframe}
\usage{
clean_strings(dat, remove_ss = TRUE)
}
\arguments{
\item{remove_ss}{logical. Default is TRUE. If TRUE superscripts are removed.
Set to FALSE if there strings of letters that end in a number that you want
to keep. Where a number falls at the end of an alphanumeric code, it will
not be interpreted as a superscript and will not be removed. However, if the
ONLY number in an alphanumeric code is at the end, the number will be seen
as a superscript.}

\item{source_data}{dataframe or tibble}
}
\value{
dataframe
}
\description{
Change factors to character strings and remove superscripts and excess white
space.
}
\examples{
clean_data <- clean_strings(source_data, remove_ss = TRUE)

}
