% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_calculations.R
\name{count_max_decimal_places}
\alias{count_max_decimal_places}
\alias{count_decimal_places_numerically}
\alias{count_decimal_places_as_string}
\title{Count number of decimal places}
\usage{
count_max_decimal_places(value)

count_decimal_places_numerically(value)

count_decimal_places_as_string(value)
}
\arguments{
\item{value}{Numeric vector.}
}
\value{
Numeric value.
}
\description{
Counts the maximum number of decimal places used for a numeric variable as
calculated using two different methods. As both methods fail when the number
of significant figures are large (but fail differently), this function
compares the two and if they return different values a warning is given.
}
\details{
count_decimal_places_numerically counts the maximum number of decimal places
used for a numeric variable. This should be used with caution as 'R stores
numbers internally in a binary representation and the exact conversion of
this representation to a decimal number may not match the exact value of the
decimal representation of the original number'.

count_decimal_places_as_string counts the maximum number of decimal places
used for a numeric variable after converting it to a character string. This
should be used with caution as the way R converts large numbers or numbers
with lots of decimal places can cause the incorrect value to be returned.
}
\examples{
test_dat <- c(1.1, 1.234)
count_max_decimal_places(test_dat) # 3dp
count_decimal_places_numerically(test_dat) # returns correct values
count_decimal_places_as_string(test_dat) # returns correct values

test_dat2 <- c(12345678891011.234)
count_max_decimal_places(test_dat2) # character count gives 1dp, while numeric count correctly gives 3dp
count_decimal_places_numerically(test_dat2) # returns correct values
count_decimal_places_as_string(test_dat2) # returns 1 because as.character() truncates number.

test_dat3 <- c(12.12345678901234567)
count_max_decimal_places(test_dat3) # character count gives 13, numeric count gives 15. Both are incorrect.
count_decimal_places_numerically(test_dat3) # returns 15, where the correct figure is 17.
count_decimal_places_as_string(test_dat3) # returns 13, where the correct figure is 17 because as.character() truncates number.

test_dat4 <- c(123456789012345678.9012345678901234567890)
count_decimal_places_as_string(test_dat4) # returns 0 because as.character() truncates number.

test_dat5 <- c(123456789012345678.234)
count_decimal_places_as_string(test_dat5) # returns 18 because of conversion to 1.234568e+18

}
\seealso{
See thread:
https://stat.ethz.ch/pipermail/r-help/2012-July/317676.html
}
