% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mismatches.R
\name{count_mismatches}
\alias{count_mismatches}
\title{Count mismatches between numbers}
\usage{
count_mismatches(number_var_1, number_var_2)
}
\arguments{
\item{number_var_1, number_var_2}{Numeric vector - both number_var vectors should be of the same length.}
}
\value{
Number of mismatches.
}
\description{
Counts the number of mismatches between two number vectors of the same length.
}
\examples{
correct_number <- c(1:4)
calculated_number <- c(1:3,5)
count_mismatches(correct_number, calculated_number)

correct_char <- letters[1:4]
created_char <- c(letters[1:3], "e")
count_mismatches(correct_char, created_char)

}
