% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datetime_for_filename.R
\name{create_datetime_for_filename}
\alias{create_datetime_for_filename}
\title{Create datetime that can be used in filename}
\usage{
create_datetime_for_filename(date_time)
}
\arguments{
\item{date_time}{output of \code{sys.time()}}
}
\value{
date and time with dash separators
}
\description{
Replaces the '-' and ':' in the date time created by \code{sys.time()} with a dash.
}
\examples{
date_time <- Sys.time()
filename_date_time <- create_datetime_for_filename(date_time)
filename_date_time

}
