% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_and_metadata.R
\name{extract_data}
\alias{extract_data}
\title{remove information above the headers}
\usage{
extract_data(dat, header_row)
}
\arguments{
\item{dat}{dataframe or tibble}

\item{header_row}{numeric. The row number on which the headers are located}
}
\value{
a list containing years and countries found in the info above the headers
}
\description{
Sometimes there will be metadata stored above the header rows. Remove this
information, leaving just the data (and potentially also footnotes) with
headers as column names. If there is no metadata above the headers, the
original data are returned.
}
\examples{
test_dat <- data.frame(
  col1 = c("Data relating to England and Wales, 2021", "", "Sex", "Female"),
  col2 = c("", "", "Age", "Under 16"))
clean_data <- extract_metadata(test_dat, 3)

}
