% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_and_metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{get year and country from info above headers}
\usage{
extract_metadata(dat, header_row, type = "not")
}
\arguments{
\item{dat}{dataframe or tibble}

\item{header_row}{numeric. The row number on which the headers are located}

\item{type}{"xlsx_cells" if data format is from xlsx_cells. Default is not
xlsx_cells}
}
\value{
a list containing years and countries found in the info above the headers
}
\description{
Extract the year and country from info above the headers. If headers are in
the first row NAs are returned.
}
\examples{
test_dat <- data.frame(
  col1 = c("Data relating to England and Wales, 2021", "", "Sex", "Female"),
  col2 = c("", "", "Age", "Under 16"))
clean_data <- extract_metadata(test_dat, 3)

}
