% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_region_names.R
\name{format_region_names}
\alias{format_region_names}
\title{Format region names}
\usage{
format_region_names(variable)
}
\arguments{
\item{variable}{Character vector of region names.}
}
\value{
Character vector. Function overwrites region variable with correct
capitalisation.
}
\description{
Corrects the capitalisation of region names
}
\examples{
test_dat <- tibble(Region = c("YORKSHIRE AND THE HUMBER", "EAST OF ENGLAND", "The shire"))
test_dat \%>\% mutate(Region = format_region_names(Region))

test_var <- c("YORKSHIRE AND THE HUMBER", "EAST OF ENGLAND", "The shire")
Region <- format_region_names(test_var)
}
