% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_usd_funcs.R
\name{gbp_to_constant_usd}
\alias{gbp_to_constant_usd}
\alias{read_exchange_rates}
\alias{read_deflators}
\alias{gbp_to_usd}
\alias{usd_to_constant_usd}
\alias{apply_exchange_rate}
\alias{apply_deflators}
\title{Convert GBP to constant USD}
\usage{
gbp_to_constant_usd(rates_filepath, deflators_filepath, gbp_data)

read_exchange_rates(rates_filepath)

read_deflators(deflators_filepath)

gbp_to_usd(exchange_rates, gbp_data)

usd_to_constant_usd(deflators, usd_data)

apply_exchange_rate(gbp_data, uk_exchange_data)

apply_deflators(usd_data, uk_deflators)
}
\arguments{
\item{rates_filepath}{Filepath of the exchange rates xlsx file - see link in
description.}

\item{deflators_filepath}{Filepath of the deflators dataframe - see link in
description.}

\item{gbp_data}{Dataframe with year and value as columns
(case is unimportant). Value should be GBP}
}
\value{
Dataframe - same as gbp_data but with constant USD instead of GBP
}
\description{
Uses exchange rates and deflators to convert GBP values to constant USD.
Exchange rate and deflator data used to convert GBP to Constant USD are from
the "data tables section" of
\href{https://www.oecd.org/dac/financing-sustainable-development/development-finance-data/}{Development finance data - OECD}.

To use this function you will first need to download
\href{https://www.oecd.org/dac/financing-sustainable-development/development-finance-data/Exchange-rates.xls}{Annual Exchange Rates for DAC Donor Countries}
and "Deflators for Resource Flows from DAC Countries" (the base year and thus
url changes with each update, so no url is given for this one).
Save both as xlsx files (they are currently downloaded as xls files).

The constant USD conversion is done using
\href{https://www.oecd.org/dac/financing-sustainable-development/development-finance-standards/informationnoteonthedacdeflators.htm}{methodology}
specified by the OECD.
The GBP values are converted with the USD exchange rate for the respective
year, and then a deflator is applied to transform the current USD values to
constant USD values. See the filename of the deflators data on
\href{https://www.oecd.org/dac/financing-sustainable-development/development-finance-data/}{Development finance data}
for the base year.
}
\section{Functions}{
\itemize{
\item \code{read_exchange_rates}: read in exchange rate xlsx file

\item \code{read_deflators}: read in deflators xlsx file

\item \code{gbp_to_usd}: convert GBP to (non-constant) USD

\item \code{usd_to_constant_usd}: convert USD to constant USD

\item \code{apply_exchange_rate}: apply US exchange rate to GBP

\item \code{apply_deflators}: apply UK deflators to USD values
}}

\examples{
{
gbp_data <- read.csv(system.file("testdata", "test_data_8-a-1.csv",
package = "SDGupdater"))
rates_filepath <- paste0(system.file(package = "SDGupdater"), "/testdata/Exchange-rates.xlsx")
deflators_filepath <- paste0(system.file(package = "SDGupdater"), "/testdata/Deflators-base-2020.xlsx")
gbp_to_constant_usd(rates_filepath, deflators_filepath, gbp_data)
}

}
\seealso{
tidy_data

get_uk_values
}
