% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_country_names.R
\name{get_all_country_names}
\alias{get_all_country_names}
\alias{multiple_country_warning}
\title{Extract country from string containing country name}
\usage{
get_all_country_names(variable)

multiple_country_warning(filename, tab, description)
}
\arguments{
\item{variable}{Character vector you want to check to see if it contains the
name of a UK country.}

\item{filename}{string giving name of the file (so warning is specific)}

\item{tab}{string giving name of the tab (so warning is specific)}

\item{description}{string describing the data that is affected (e.g. which disaggregation)}
}
\value{
Character vector giving name of country or NA.

a warning
}
\description{
Identifies whether a string contains the name of one or more of the UK
countries and if it does, it gives the names of that country as a string,
otherwise it returns NA.
}
\section{Functions}{
\itemize{
\item \code{multiple_country_warning}: Extract country from string containing country name
}}

\examples{
test_dat <- c("England is a country of the UK", "this does not contain a
country name") get_all_country_names(test_dat)

multiple_country_warning("foo.xlsx", "Tab 1", "sex_by_age disaggregation")

}
