% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_info_above_column_names.R
\name{get_all_years}
\alias{get_all_years}
\alias{multiple_year_warning}
\title{Extract years from strings}
\usage{
get_all_years(variable)

multiple_year_warning(filename, tab, description)
}
\arguments{
\item{variable}{Character vector.}
}
\value{
Character vector giving year or NA.
}
\description{
Identifies whether a string contains a number that looks like a year from the
20th or 21st century. If it does, it returns the year as a string, otherwise
it returns NA. If it contains more than one year, all years are returned as a
comma separated string
}
\examples{
test_dat <- c("a date: 2005", "1995 6", "1993 end", "a1995", "1995a", "a
date:1995, another date: 2012", "not a date: 475", "also not a date: 1234",
"abc 12345",  "34200943", "190932")#' get_all_country_names(test_dat)
get_all_years(test_dat)

}
