% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_characters_after_dot}
\alias{get_characters_after_dot}
\title{Get all charcters following a dot}
\usage{
get_characters_after_dot(string)
}
\arguments{
\item{variable}{Character string.}
}
\value{
Character string.
}
\description{
Get all charcters following the last dot in a character string
}
\examples{
get_characters_after_dot("ab.cde")
get_characters_after_dot("ab.cde.fgh")

}
