% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_column.R
\name{get_column_index-deprecated}
\alias{get_column_index-deprecated}
\alias{get_column_index}
\title{Get index of column name where a pattern is matched}
\usage{
get_column_index(dat, pattern)

get_column_index(dat, pattern)
}
\arguments{
\item{dat}{vector.}

\item{pattern}{vector of strings to look for}
}
\value{
indices of all columns matching the pattern(c)
}
\description{
Returns all indices of a vector of column names where a pattern is matched.
If multiple patterns are passed, only indices where all patterns are matched
will be  returned.
}
\section{\code{get_column_index}}{

For \code{get_column_index}, use \code{\link{get_indices}}.
}

\examples{
test_data <- c("numbers_1_4", "numbers_10_14", "rates_1_4")
indices <- get_indices(test_data, c("numbers", "1"))

}
\seealso{
\code{\link{SDGupdater-deprecated}}
}
\keyword{internal}
