% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_column.R
\name{get_column_index}
\alias{get_column_index}
\title{Get index of column names matching a pattern}
\usage{
get_column_index(dat, pattern)
}
\arguments{
\item{dat}{dataframe or tibble.}
}
\value{
index of columns matching the pattern
}
\description{
Use patterns that are expected to always appear in a column name to rename
the column. This provides some future-proofing for changes to column names in
future input data. Columns will only be renamed if exactly one column is identified.
If this is not the case a warning will be returned, and the column will remain
with it's original name
}
\examples{
test_data <- data.frame("numbers_1_4" = c(10, 5),
                        "numbers_10_14" = c(1, 3),
                        "rates_1_4" = c(4, 7))
indices <- get_column_index(test_data, c("numbers", "1"))

}
