% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fourth_from_last_character.R
\name{get_fourth_from_last_character}
\alias{get_fourth_from_last_character}
\title{Get fourth from last character}
\usage{
get_fourth_from_last_character(variable)
}
\arguments{
\item{variable}{Character string.}
}
\value{
Character.
}
\description{
Get the fourth from last character of a string
}
\examples{
get_fourth_from_last_character("abcde")

}
