% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_column.R
\name{get_indices}
\alias{get_indices}
\title{Get indices of vector where a pattern is matched}
\usage{
get_indices(dat, pattern)
}
\arguments{
\item{dat}{vector.}

\item{pattern}{vector of strings to look for}
}
\value{
indices of all columns matching the pattern(c)
}
\description{
Returns all indices of a vector where a pattern is matched. If multiple
patterns are passed, only indices where all patterns are matched will be
returned.
}
\examples{
test_data <- c("numbers_1_4", "numbers_10_14", "rates_1_4")
indices <- get_indices(test_data, c("numbers", "1"))

}
