% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_cells.R
\name{get_info_cells}
\alias{get_info_cells}
\title{Isolate cells above column headings}
\usage{
get_info_cells(dat, first_header_row, type = "xlsx_cells")
}
\arguments{
\item{dat}{Dataframe imported using xlsx_cells where there are header cells
above the column names containing info on date and year.}

\item{first_header_row}{Numeric value. The row number in the excel file
containing the first set of column headings.}
}
\value{
Tibble containing all year and country info.
}
\description{
Gets year and country info from the rows above the column headings.
Can be used for data imported using xlsx_cells, without first changing it to
a more standard format. Alternatively it can be used on a more standard
dataframe, for example one imported with read.xlsx.
}
\examples{
first_header_row <- 4
test_dat <- data.frame(row = c(1:5), character = c("data for 2017 ", "England and", "Wales", "impala", "elephant"))
get_info_cells(test_dat, first_header_row)

}
