% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modal_values.R
\name{get_modal_values}
\alias{get_modal_values}
\title{Get modal values}
\usage{
get_modal_values(values)
}
\arguments{
\item{values}{Numeric or character vector.}
}
\value{
Vector of same type as values.
}
\description{
Finds which value occurs most frequently. If there is more than one mode, all
will be returned
}
\seealso{
From thread:
https://stackoverflow.com/questions/2547402/how-to-find-the-statistical-mode
}
