% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_penultimate_character}
\alias{get_penultimate_character}
\title{Get penultimate character}
\usage{
get_penultimate_character(variable)
}
\arguments{
\item{variable}{Character string.}
}
\value{
Character.
}
\description{
Get the second to last character of a string
}
\examples{
get_penultimate_character("abcde")

}
