% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_type1_data}
\alias{get_type1_data}
\title{read type 1 data}
\usage{
get_type1_data(header_row, filename, tabname, na.strings = "")
}
\arguments{
\item{header_row}{numeric. The row number on which the headers are located}

\item{filename}{character string. name of an excel file including the extension.
This is pasted to input_folder to create the filepath}

\item{tabname}{character string. The name of the sheet, exactly as it is in the
excel file}

\item{na.strings}{A character vector of strings which are to be interpreted as NA.
Defaults to "". Note that for csv files the test happens after white space is
stripped from the input, so na.strings values may need their own white space
stripped in advance.}
}
\value{
dataframe
}
\description{
Read in a single sheet of excel data, where there is only one row of headers.
Data are read in slightly differently depending on whether headers are in the
first row or not
}
\examples{
source_data <- get_type1_data(header_row, filename, tabname)

}
