% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_usd_funcs.R
\name{get_uk_values}
\alias{get_uk_values}
\title{Select UK deflator/exchange rate values}
\usage{
get_uk_values(dat, value_name)
}
\arguments{
\item{dat}{Exchange rate or deflators data frame (link in description)}

\item{value_name}{The new name of the value column}
}
\value{
dataframe - uk rows of dat
}
\description{
Selects UK exchange rate and deflators data from
\url{https://www.oecd.org/dac/financing-sustainable-development/development-finance-data/}
so that it can be used in conversions.
}
\seealso{
gbp_to_constant_usd
}
