% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_absent_packages.R
\name{install_absent_packages}
\alias{install_absent_packages}
\title{Install packages that are not already installed}
\usage{
install_absent_packages(packages)
}
\arguments{
\item{packages.}{A vector of packages you want to install if they are not already}
}
\description{
Takes a vector of package names that you need and installs
any packages that are not already installed. Returns a message to say all are
already installed, or installs any absent packages, with the usual
install.packages messages.
}
\examples{
install_absent_packages(c("dpyr", "tidyr"))

}
