% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_years.R,
%   R/list_into_vector_of_strings.R
\name{list_into_vector_of_strings}
\alias{list_into_vector_of_strings}
\title{Turn a list of character vectors/matrices into a vector of character strings}
\usage{
list_into_vector_of_strings(input_list)

list_into_vector_of_strings(input_list)
}
\arguments{
\item{input_list}{List of character vectors or matrices.}
}
\value{
Vector with one string for each component of the list. If the
input_list is a matrix, the order is first down rows then across columns.

Vector with one string for each component of the list. If the
input_list is a matrix, the order is first down rows then across columns.
}
\description{
Turns a list of vectors/matrices that hold only character strings into a
single vector of character strings, separated by ", ". The elements that make
up each component of the input list are pasted into a single string. Each
element of the new vector is formed from one component of the input list.

Turns a list of vectors/matrices that hold only character strings into a
single vector of character strings, separated by ", ". The elements that make
up each component of the input list are pasted into a single string. Each
element of the new vector is formed from one component of the input list.
}
\examples{
matrix_1 <- as.matrix(c("England", "UK"))
matrix_2 <- as.matrix(c("Wales"))
matrix_3 <- as.matrix(c("NA"))
matrix_list <- list(matrix_1, matrix_2, matrix_3)
list_into_vector_of_strings(matrix_list)

matrix_1 <- as.matrix(c("England", "UK"))
matrix_2 <- as.matrix(c("Wales"))
matrix_3 <- as.matrix(c("NA"))
matrix_list <- list(matrix_1, matrix_2, matrix_3)
list_into_vector_of_strings(matrix_list)

}
