% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_column.R
\name{pinpoint_indices}
\alias{pinpoint_indices}
\title{Get indices where a set of pattern conditions are met}
\usage{
pinpoint_indices(dat, primary, alternate, not_pattern)
}
\arguments{
\item{dat}{vector.}

\item{primary}{first pattern or patterns to look for. Alternate will be
ignored if this is found}

\item{alternate}{pattern(s) to look for if nothing is returned for primary}

\item{not_pattern}{exclude any indices where this pattern is matched}
}
\value{
vector of all indices where conditions are met
}
\description{
Allows the user to specify a set of pattern conditions and returns all
inidces of a vector where those conditions are met.
}
\details{
If multiple patterns are given for an argument, the condition will only be
met if ALL patterns in that argument are matched.
}
\examples{
test_data <- c("numbers_1_4", "numbers_10_14", "rates_1_4" = c(4, 7))
indices <- get_indices(test_data, c("numbers", "1"))

}
