% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_clean_data.R
\name{remove_blanks_and_info_cells}
\alias{remove_blanks_and_info_cells}
\title{Remove blanks and cells above the column headings}
\usage{
remove_blanks_and_info_cells(dat, first_header_row)
}
\arguments{
\item{dat}{Tibble originaly imported using xlsx_cells.}

\item{first_header_row}{Numeric. Row number of first row in the xlsx file
containing column names.}
}
\value{
Tibble.
}
\description{
Remove blanks and cells above the column headings to clean xlsx_cells data.
Do this before using behead()
}
\examples{
test_dat <- dplyr::tibble(row = c(1:6),
                   character = c("2017", NA, "England", "Rates", NA, NA),
                   is_blank = c(FALSE, TRUE, FALSE, FALSE, TRUE, FALSE),
                   data_type = c("character", "blank", "character", "character", "blank", "numeric"),
                   numeric = c(NA, NA, NA, NA, NA, 20))
remove_blanks_and_info_cells(test_dat, 4)
}
