% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_footnotes.R
\name{remove_footnotes}
\alias{remove_footnotes}
\title{Remove footnotes below table}
\usage{
remove_footnotes(data, check_columns)
}
\arguments{
\item{data}{dataframe or tibble.}

\item{check_columns}{numeric. The maximum index in which you expect there to be footnotes}
}
\value{
data frame with footnote, and all-NA rows removed
}
\description{
Assuming footnotes are in the first few columns, this function removes those
rows, leaving only rows with data.

remove_footnotes starts at the bottom of the dataframe and stops looking
for footnotes as soon as a row that looks like a non-footnote row is
encountered.

Footnotes must be in one or more of the first x columns,
where x = check_columns.
If the data are likely to have non-footnote (real) data in the first x
column(s) and NAs in all other columns DO NOT use this function as it will
likely remove more than just footnotes
}
\examples{

test_data <- data.frame("A" = c(10, 5, "footnote:"),
                        "B" = c(1, 3, "this is a footnote"),
                        "C" = c(4, 7, NA),
                        "D" = c(5, 9, NA))
remove_footnotes(test_data, 2)
                        
}
