% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_superscripts.R
\name{remove_superscripts}
\alias{remove_superscripts}
\alias{remove_real_superscripts}
\alias{remove_multiple_superscripts}
\alias{remove_false_superscripts}
\alias{could_contain_superscript}
\title{Remove superscripts}
\usage{
remove_superscripts(variable)

remove_real_superscripts(variable)

remove_multiple_superscripts(variable)

remove_false_superscripts(variable)

could_contain_superscript(variable)
}
\arguments{
\item{variable}{Character vector you want superscripts removed from.}
}
\value{
Character vector.
}
\description{
Remove superscripts, whether they are read as actual
superscripts or just as numbers. This function assumes that superscripts
never follow:
\itemize{
\item a space
\item a number (so superscripts over 9 are not recognised)
\item a string consisting of just one other character
\item a 'word' consisting of both numbers and letters
\item £ or $
}
}
\examples{
character <- c("Wales1","Rates?", "Numbers1,2")
remove_superscripts(character)
remove_real_superscripts(character)
remove_multiple_superscripts(character)
remove_false_superscripts(character)
}
