% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_column.R
\name{rename_column}
\alias{rename_column}
\title{Rename columns based on patterns in current name}
\usage{
rename_column(dat, primary, alternate, not_pattern, new_name)
}
\arguments{
\item{dat}{dataframe or tibble.}

\item{primary}{vector of expected patterns in the column to be renamed}

\item{alternate}{alternative pattern to pattern. Only used if pattern is unmatched}

\item{not_pattern}{vector of patterns that must not be present in the column to be renamed}

\item{new_name}{new name for the column}
}
\value{
dat with columns renamed
}
\description{
Use patterns that are expected to always appear in a column name to rename
the column. This provides some future-proofing for changes to column names in
future input data. Columns will only be renamed if exactly one column is identified.
If this is not the case a warning will be returned, and the column will remain
with it's original name
}
\examples{

test_data <- data.frame("numbers_1_4" = c(10, 5),
                        "numbers_10_14" = c(1, 3),
                        "rates_1_4" = c(4, 7))
renamed <- rename_column(dat = test_data,
                        primary = c("number", "1", "4"),
                        alternate = c("death", "1", "4"),
                        not_pattern = c("10"),
                        new_name = "number_deaths_1_to_4")
                        
}
