% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_usd_funcs.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy exchange rate and deflators data}
\usage{
tidy_data(dat, heading_row)
}
\arguments{
\item{dat}{Exchange rate or deflators data frame (link in description)}

\item{heading_row}{Numeric. The row on which headings are given in the xlsx}
}
\value{
dataframe - gbp_data with an additional column giving value as
constant USD
}
\description{
Tidies exchange rate and deflators data from
\url{https://www.oecd.org/dac/financing-sustainable-development/development-finance-data/}
so that it can be used in conversions.
}
\seealso{
gbp_to_constant_usd
}
