% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_info_above_column_names.R
\name{unique_to_string}
\alias{unique_to_string}
\title{Unique entries to string}
\usage{
unique_to_string(variable)
}
\arguments{
\item{variable}{Vector of character strings.}
}
\value{
Single character string of unique entries, not including NAs.
}
\description{
Turns a vector that hold only character strings into a single character
string, with values separated by separated by ", ". NAs are ignored.
}
\examples{
Year <- c("2018", "2015, 2016", NA)
Country <- c(NA, NA, "United Kingdom")
unique_to_string(Year)
unique_to_string(Country)

}
