% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_messages.R
\name{wrong_tab_message}
\alias{wrong_tab_message}
\title{wrong tab warning based on key words}
\usage{
wrong_tab_message(tab_name, desired_data)
}
\arguments{
\item{tab_name}{String}

\item{desired_data}{String describing desired data (e.g. table description).}
}
\value{
String
}
\description{
Message for when expected key words are not found
}
\examples{
warning(wrong_tab_message("Table 2", "territorial greenhouse gas emissions by source sector"))
}
