% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_user_interaction.R
\name{ask_user_for_filename}
\alias{ask_user_for_filename}
\title{Get filename}
\usage{
ask_user_for_filename(file_path = ".")
}
\arguments{
\item{file_path}{Character string. If left blank, the working directory is used.}
}
\value{
Character string.
}
\description{
Ask the user to select the file they want to work with. All files in the current directory are given as options.
}
\examples{
ask_user_for_filename()
}
