% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_years.R, R/get_info_cells.R
\name{get_info_cells}
\alias{get_info_cells}
\title{Isolate cells above column headings}
\usage{
get_info_cells(dat, first_header_row)

get_info_cells(dat, first_header_row)
}
\arguments{
\item{dat}{Dataframe imported using xlsx_cells where there are header cells
above the column names containing info on date and year.}

\item{first_header_row}{Numeric value. The row number in the excel file
containing the first set of column headings.}
}
\value{
Tibble containing all year and country info.

Tibble containing all year and country info.
}
\description{
Gets year and country info from the rows above the column headings in a
dataframe created using xlsx_cells

Gets year and country info from the rows above the column headings in a
dataframe created using xlsx_cells
}
\details{
Assumes that the cell containing the year ends with the year

Assumes that the cell containing the year ends with the year
}
\examples{
first_header_row <- 4 test_dat <- data.frame(row = c(1:5), character =
c("data for 2017 ", "England and", "this is Wales", "impala", "elephant"))
get_info_cells(test_dat, first_header_row)

first_header_row <- 4 test_dat <- data.frame(row = c(1:5), character =
c("data for 2017 ", "England and", "this is Wales", "impala", "elephant"))
get_info_cells(test_dat, first_header_row)

}
