% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_superscripts.R
\name{remove_superscripts}
\alias{remove_superscripts}
\alias{remove_real_superscripts}
\alias{remove_multiple_superscripts}
\alias{remove_false_superscripts}
\alias{could_contain_superscript}
\title{Remove superscripts}
\usage{
remove_superscripts(variable)

remove_real_superscripts(variable)

remove_multiple_superscripts(variable)

remove_false_superscripts(variable)

could_contain_superscript(variable)
}
\arguments{
\item{variable}{Character vector you want superscripts removed from.}
}
\value{
Character vector for all but \code{could_contain_superscripts}, which
returns a logical vector.
}
\description{
Remove superscripts, whether they are read as actual
superscripts or just as numbers (which are referred to as 'false
superscripts').
}
\details{
This function assumes that superscripts never follow:
\itemize{
\item a space
\item a number (superscripts over 9 are not recognised)
\item a string consisting of just one other character
\item a 'word' consisting of both numbers and letters
\item a pound or dollar symbol
\item a string starting with a number or a single letter
}

If it becomes apparent that these scenarios need to be covered, the function
will need updating.
}
\examples{
character <- c("Wales1","Rates?", "Numbers1,2")
remove_superscripts(character)
remove_real_superscripts(character)
remove_multiple_superscripts(character)
remove_false_superscripts(character)
}
